# shared/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactRefDto": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "artifact_type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "checksum": {
            "title": "Checksum",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "relative_path": {
            "title": "Relative Path",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Id"
          }
        },
        "required": [
          "artifact_id",
          "run_id",
          "artifact_type",
          "relative_path",
          "created_at",
          "checksum"
        ],
        "title": "ArtifactRefDto",
        "type": "object"
      },
      "ArtifactType": {
        "enum": [
          "json",
          "yaml",
          "markdown",
          "text_log",
          "command_log",
          "patch",
          "diff",
          "report"
        ],
        "title": "ArtifactType",
        "type": "string"
      },
...
      "AuthoritativeRunMutationResultDto": {
        "additionalProperties": false,
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRefDto"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "event_sequence": {
            "minimum": 1.0,
            "title": "Event Sequence",
            "type": "integer"
          },
          "graph_thread_id": {
            "title": "Graph Thread Id",
            "type": "string"
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "state_version": {
            "minimum": 1.0,
            "title": "State Version",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status",
          "state_version",
          "event_sequence",
          "graph_thread_id"
        ],
        "title": "AuthoritativeRunMutationResultDto",
        "type": "object"
      },
...
      "BaselineTargetInventoryResponse": {
        "additionalProperties": false,
        "properties": {
          "angular_json_present": {
            "title": "Angular Json Present",
            "type": "boolean"
          },
          "event_sequence": {
            "title": "Event Sequence",
            "type": "integer"
          },
          "package_json_checksum": {
            "title": "Package Json Checksum",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "state_version": {
            "title": "State Version",
            "type": "integer"
          },
          "targets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "targets",
          "package_json_checksum",
          "angular_json_present",
          "state_version",
          "event_sequence"
        ],
        "title": "BaselineTargetInventoryResponse",
        "type": "object"
      },
      "BaselineValidationRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "expected_state_version": {
            "minimum": 1.0,
            "title": "Expected State Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "prerequisite_artifact_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisite Artifact Ids",
            "type": "array"
          }
        },
        "required": [
          "expected_state_version",
          "idempotency_key",
          "actor"
        ],
        "title": "BaselineValidationRequest",
        "type": "object"
      },
      "BaselineValidationResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact_checksums": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Artifact Checksums",
            "type": "object"
          },
          "artifact_ids": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Ids",
            "type": "array"
          },
          "baseline_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Checksum"
          },
          "event_sequence": {
            "title": "Event Sequence",
            "type": "integer"
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "build",
              "test",
              "lint"
            ],
            "title": "Kind",
            "type": "string"
          },
          "parser_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Summary"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Results",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "state_version": {
            "title": "State Version",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "targets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Targets",
            "type": "array"
          },
          "validation_id": {
            "title": "Validation Id",
            "type": "string"
          }
        },
        "required": [
          "validation_id",
          "run_id",
          "kind",
          "status",
          "targets",
          "results",
          "state_version",
          "event_sequence"
        ],
        "title": "BaselineValidationResponse",
        "type": "object"
      },
...
      "CreateAuthoritativeRunRequestDto": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "artifact_set_checksum": {
            "minLength": 1,
            "title": "Artifact Set Checksum",
            "type": "string"
          },
          "client_constraints": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Client Constraints",
            "type": "object"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "input_checksum": {
            "minLength": 1,
            "title": "Input Checksum",
            "type": "string"
          },
          "preflight_id": {
            "minLength": 1,
            "title": "Preflight Id",
            "type": "string"
          },
          "pricing_snapshot": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Pricing Snapshot",
            "type": "object"
          }
        },
        "required": [
          "preflight_id",
          "input_checksum",
          "artifact_set_checksum",
          "idempotency_key",
          "actor"
        ],
        "title": "CreateAuthoritativeRunRequestDto",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "CREATED",
          "SOURCE_VALIDATION_RUNNING",
          "SOURCE_VALIDATED",
          "WORKSPACE_CLASSIFICATION_RUNNING",
          "BASELINE_RUNNING",
          "BASELINE_QUALIFIED",
          "CLIENT_CONSTRAINTS_CAPTURED",
          "ELIGIBILITY_RUNNING",
          "ANALYSIS_RUNNING",
          "WAITING_ANALYSIS_APPROVAL",
          "PLANNING_RUNNING",
          "WAITING_PLAN_APPROVAL",
          "STAGE_CREATED",
          "TOOLCHAIN_PROFILE_SELECTED",
          "SANDBOX_READY",
          "DEPENDENCY_AUDITED",
          "TRANSFORMATION_RUNNING",
          "STATIC_SYMBOL_CHECK_RUNNING",
          "VALIDATION_RUNNING",
          "REPAIR_RUNNING",
          "WAITING_REPAIR_APPROVAL",
          "REVIEW_READY",
          "STAGE_COMMITTED",
          "STAGE_ROLLED_BACK",
          "REPORT_RUNNING",
          "DELIVERY_RUNNING",
          "PAUSE_REQUESTED",
          "PAUSED",
          "RESUMING",
          "CANCEL_REQUESTED",
          "CANCELLING",
          "TIMED_OUT",
          "WORKER_LOST",
          "RECOVERY_RUNNING",
          "ORPHANED",
          "CLEANUP_RUNNING",
          "CLEANUP_FAILED",
          "RUNNING",
          "WAITING",
          "CANCELLED",
          "COMPLETED",
          "FAILED",
          "DIAGNOSTIC_HOLD"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend execution-authority API for controlled frontend migration workflows.",
    "title": "AI Frontend Migration Factory API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthoritativeRunRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthoritativeRunMutationResultDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/baseline/builds": {
      "post": {
        "operationId": "execute_builds_api_v1_runs__run_id__baseline_builds_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Builds",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/baseline/lint": {
      "post": {
        "operationId": "execute_lint_api_v1_runs__run_id__baseline_lint_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Lint",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/baseline/targets": {
      "get": {
        "operationId": "get_targets_api_v1_runs__run_id__baseline_targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineTargetInventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Targets",
        "tags": [
          "baseline-validation"
        ]
      }
    },
    "/api/v1/runs/{run_id}/baseline/tests": {
      "post": {
        "operationId": "execute_tests_api_v1_runs__run_id__baseline_tests_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Tests",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/baseline/{kind}": {
      "get": {
        "operationId": "get_validation_api_v1_runs__run_id__baseline__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "build",
                "test",
                "lint"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Validation",
        "tags": [
          "baseline-validation"
        ]
      }
    },
    "/api/v1/runs/{run_id}/baseline/{kind}/cancel": {
      "post": {
        "operationId": "cancel_validation_api_v1_runs__run_id__baseline__kind__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "build",
                "test",
                "lint"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Validation",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/runs": {
      "post": {
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthoritativeRunRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthoritativeRunMutationResultDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/runs/{run_id}/baseline/builds": {
      "post": {
        "operationId": "execute_builds_runs__run_id__baseline_builds_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Builds",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/runs/{run_id}/baseline/lint": {
      "post": {
        "operationId": "execute_lint_runs__run_id__baseline_lint_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Lint",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/runs/{run_id}/baseline/targets": {
      "get": {
        "operationId": "get_targets_runs__run_id__baseline_targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineTargetInventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Targets",
        "tags": [
          "baseline-validation"
        ]
      }
    },
    "/runs/{run_id}/baseline/tests": {
      "post": {
        "operationId": "execute_tests_runs__run_id__baseline_tests_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Tests",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
    "/runs/{run_id}/baseline/{kind}": {
      "get": {
        "operationId": "get_validation_runs__run_id__baseline__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "build",
                "test",
                "lint"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Validation",
        "tags": [
          "baseline-validation"
        ]
      }
    },
    "/runs/{run_id}/baseline/{kind}/cancel": {
      "post": {
        "operationId": "cancel_validation_runs__run_id__baseline__kind__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "build",
                "test",
                "lint"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Validation",
        "tags": [
          "baseline-validation"
        ]
      }
    },
...
  }
}